remove unnecessary timestamp parsing
The LastModified timestamp attribute is of type System.DateTimeOffset. Therefore, the parsing of LastModified is not necessary. The parsing also failed on environments with a CultureInfo different from Invariant.

Since the goal seems to be to have a System.DateTime type for LastModified instead of System.DateTimeOffset, the UtcDateTime representation of LastModified was chosen as per https://learn.microsoft.com/en-us/dotnet/standard/datetime/converting-between-datetime-and-offset#conversions-from-datetimeoffset-to-datetime